requirements.txt
# Django and core dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.2.0
django-crispy-forms==2.1
djangorestframework-simplejwt==5.2.2


# Database
psycopg2-binary==2.9.9   # PostgreSQL (remove if using SQLite/MySQL)
dj-database-url==2.1.0

# Deployment
gunicorn==21.2.0         # For production WSGI server
whitenoise==6.6.0        # Serve static files easily

# Environment variables
python-dotenv==1.0.0

# Optional: Cloud media storage
cloudinary==1.41.0

# Utilities
asgiref==3.7.2
sqlparse==0.4.4
